refactor(kotlin native, api): split declarations of kotlin native api module
TODO make signal module depends on binding module
