+++ kotlin-native/api/chore/build.gradle.kts
@@ -1,0 +1,28 @@
+import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+
+plugins {
+    alias(libs.plugins.kotlin.multiplatform.conventions)
+    alias(libs.plugins.kotlin.styles.conventions)
+}
+
+kotlin {
+    compilerOptions {
+        explicitApi()
+        optIn.addAll(
+            "kotlinx.cinterop.ExperimentalForeignApi",
+            "kotlin.experimental.ExperimentalNativeApi",
+            "io.github.kingg22.godot.api.ExperimentalGodotApi",
+            "io.github.kingg22.godot.api.ExperimentalGodotKotlin",
+        )
+    }
+
+    @OptIn(ExperimentalKotlinGradlePluginApi::class)
+    dependencies {
+        api(libs.jetbrains.annotations)
+    }
+
+    applyDefaultHierarchyTemplate()
+
+    linuxX64()
+    // mingwX64()
+}
