+++ kotlin-native/api/chore/src/nativeMain/kotlin/io/github/kingg22/godot/api/GodotEnum.kt
