+++ kotlin-native/api/signal/src/nativeMain/kotlin/io/github/kingg22/godot/api/signal/Signal1.kt
