+++ kotlin-native/api/signal/src/nativeTest/kotlin/io/github/kingg22/godot/api/signal/SignalParameterDescriptorTest.kt
