+++ kotlin-native/api/testing/src/nativeMain/kotlin/io/github/kingg22/godot/api/VariantTypeTestData.kt
