+++ kotlin-native/api/utils/src/nativeMain/kotlin/io/github/kingg22/godot/api/builtin/internal/ConvertToVariant.kt
