+++ settings.gradle.kts
@@ -73,6 +73,11 @@ include(
     "kotlin-native:ffi",
     "kotlin-native:annotations",
     "kotlin-native:api",
+    "kotlin-native:api:chore",
+    "kotlin-native:api:generated",
+    "kotlin-native:api:signal",
+    "kotlin-native:api:testing",
+    "kotlin-native:api:utils",
     "kotlin-native:binding",
     "kotlin-native:runtime",
     "kotlin-native:sample",
